requirements.txt
# requirements.txt

ccxt>=2.0.0
pandas>=1.3.0
numpy>=1.20.0
dash>=2.0.0
dash-bootstrap-components>=1.0.0
plotly>=5.3.0
python-telegram-bot>=20.0
python-dotenv>=0.19.0
talib-binary>=0.4.24
requests>=2.26.0
websocket-client>=1.2.1
fastapi>=0.68.0
uvicorn>=0.15.0
psutil>=5.8.0
python-jose>=3.3.0
pytest>=6.2.5
black>=21.9b0
flake8>=3.9.2
isort>=5.9.3
python-dotenv
